fix: timeline month 넘어가지 않는 이슈 컴포넌트 분리로 해결
